+++ namespaces_test.go
@@ -212,8 +212,8 @@ func TestGetNamespace(t *testing.T) {
 		WebURL:                      "https://gitlab.example.com/groups/group1",
 		MembersCountWithDescendants: 2,
 		BillableMembersCount:        2,
-		MaxSeatsUsed:                Ptr(0),
-		SeatsInUse:                  Ptr(0),
+		MaxSeatsUsed:                Ptr(int64(0)),
+		SeatsInUse:                  Ptr(int64(0)),
 		Plan:                        "default",
 		TrialEndsOn:                 nil,
 		Trial:                       false,
@@ -239,7 +239,7 @@ func TestNamespaceExists(t *testing.T) {
 	})
 
 	opt := &NamespaceExistsOptions{
-		ParentID: Ptr(1),
+		ParentID: Ptr(int64(1)),
 	}
 	exists, _, err := client.Namespaces.NamespaceExists("my-group", opt)
 	if err != nil {
@@ -297,8 +297,8 @@ func TestSearchNamespace(t *testing.T) {
 			WebURL:                      "https://gitlab.example.com/groups/twitter",
 			MembersCountWithDescendants: 2,
 			BillableMembersCount:        2,
-			MaxSeatsUsed:                Ptr(0),
-			SeatsInUse:                  Ptr(0),
+			MaxSeatsUsed:                Ptr(int64(0)),
+			SeatsInUse:                  Ptr(int64(0)),
 			Plan:                        "default",
 			TrialEndsOn:                 nil,
 			Trial:                       false,
